requirements.txt
asgiref==3.2.7
autopep8==1.5.2
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.5
django-filter==2.2.0
django-filters==0.2.1
django-silk==4.0.1
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.3
Fuzzy==1.2.2
gprof2dot==2019.11.30
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
